+++ semana11/aula3/exDaAula/ex2.ts
