+++ semana11/aula3/package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "tsc && node build/ex3.js --target es6"
+    "start": "tsc && node build/ex1.js"
   },
   "author": "",
   "license": "ISC",
