+++ semana11/aula3/src/ex1.ts
@@ -1,17 +1,12 @@
-const myPromise = new Promise((resolve, reject) => {
-    setTimeout(() => {
-        console.log("acabou");
-        resolve();
-    }, 2000)
-});
-/*
-myPromise.then(() => {
-    console.log("Promise finalizada com sucesso");
+import axios, { AxiosResponse } from 'axios';
 
-})
-*/
+const fs = require('fs');
+const fileName1 = require("1.txt")
+const fileName2 = require("2.txt")
+const fileName3 = require("3.txt")
+const fileName4 = require("4.txt")
+const fileName5 = require("5.txt")
 
-const teste = async () => {
-    await myPromise;
-    console.log("Promise finalizada com sucesso")
-};
+//fs.readdir(fileName1, function(files))
+
+console.log(fileName1)
