+++ semana11/aula3/src/textos/1.txt
@@ -1,0 +1,1 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. At augue eget arcu dictum varius duis at. Libero justo laoreet sit amet cursus sit amet dictum. Dictum non consectetur a erat nam at lectus. Montes nascetur ridiculus mus mauris vitae ultricies leo.
