+++ semana11/aula3/src/textos/2.txt
@@ -1,0 +1,1 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Id diam maecenas ultricies mi eget mauris pharetra. Sollicitudin ac orci phasellus egestas tellus rutrum tellus. Pulvinar mattis nunc sed blandit. Ultrices dui sapien eget mi proin sed libero enim.
