+++ semana11/aula3/src/textos/3.txt
@@ -1,0 +1,1 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Varius sit amet mattis vulputate enim nulla aliquet porttitor. Et netus et malesuada fames ac turpis egestas. Dis parturient montes nascetur ridiculus. Et egestas quis ipsum suspendisse ultrices.
