+++ semana11/aula3/src/textos/4.txt
@@ -1,0 +1,1 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. At imperdiet dui accumsan sit amet nulla facilisi morbi tempus. Massa id neque aliquam vestibulum morbi blandit. Ac tortor dignissim convallis aenean et tortor at risus viverra. Gravida neque convallis a cras semper.
