+++ semana11/aula3/src/textos/5.txt
@@ -1,0 +1,1 @@
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Phasellus vestibulum lorem sed risus ultricies tristique nulla. Rhoncus aenean vel elit scelerisque mauris. Pulvinar pellentesque habitant morbi tristique senectus et netus et. Duis tristique sollicitudin nibh sit amet.
